SetForm支持多参数
